Add two patch candidates to fix FIPS problems in waf.
The md5 hash algorithm is known to be insecure, and systems with the
FIPS flag set disable it. Some versions of Python appear to enforce this
restriction, and disable the md5() function in the hashlib package.

Unfortunately, waflib uses that function.

This commit includes two possible patches to address this problem.

The first is the one applied by the waf maintainer to a later version of
waf than the one we are based on.  It's a fairly ugly hack, which
aliases the sha1 hash as md5.  This will cause problems if waf ever
tries to compare against an actual md5 hash, but I'm not sure if it ever
does, and this patch has been in waf since 2018, so it probably hasn't
screwed anything up too badly.

The second option uses the relatively undocumented
'notforsecurity=False' option to md5(), which, at least in some versions
of hashlib, bypasses the FIPS lockout.  This is much cleaner, but it may
not be universally supported, and will break the build on versions of
Python which do not support it.

The two patches are compatible, and may both be applied if desired.  To
apply a patch, execute 'waf --help'; this will trigger the waflib
library to be unpacked.  The patch(es) can then be applied to the files
within.

One or both of these patches should probably be applied to the 'waf'
self-extracting package itself, after proper review.