Rebuild the waf self-extractor with a fix for the FIPS error.
Resubmission of 9f6f775f8f7628d912ab76ce99a502658ccde385.

Apply a fix from the waf maintainers to work around the disabling
  of the md5 hash function in FIPS-enabled systems.  This function
  is used internally by waf.

This isn't a great fix - it pulls in sha1 and aliases md5 to it.
But it's what they did, and it seems to work.

The issue was https://gitlab.com/ita1024/waf/issues/2213.
The commits there were bd15a9dc and 1c7be35d.

This patch remains in build/waf_fips.patch.
There is an alternative and ideally better patch in
waf_fips_alternative.patch, but not all python/libssl versions support
 the necessary 'usedforsecurity' flag.

Also add instructions for repacking the waf self-extractor.